+++ StockScraper/DLL/Reuters_Financials_Institutions.cs
@@ -14,7 +14,7 @@ namespace DLL
     
     public partial class reuters_Financials_Institutions
     {
-        public long Reuters_FinancialsInstitutions_Id { get; set; }
+        public int Reuters_FinancialsInstitutions_Id { get; set; }
         public int Stock_Id { get; set; }
         public int job_run_id { get; set; }
         public string EffectiveDate { get; set; }
