+++ StockScraper/DLL/Reuters_Financials_SalesEstimates.cs
@@ -14,7 +14,7 @@ namespace DLL
     
     public partial class reuters_Financials_SalesEstimates
     {
-        public long Reuters_FinancialsSalesEstimates_Id { get; set; }
+        public int Reuters_FinancialsSalesEstimates_Id { get; set; }
         public int Stock_Id { get; set; }
         public int job_run_id { get; set; }
         public string EffectiveDate { get; set; }
