+++ StockScraper/DLL/Reuters_Financials_ValuationRatios.cs
@@ -14,7 +14,7 @@ namespace DLL
     
     public partial class reuters_Financials_ValuationRatios
     {
-        public long Reuters_FinancialsValuationRatio_Id { get; set; }
+        public int Reuters_FinancialsValuationRatio_Id { get; set; }
         public int Stock_Id { get; set; }
         public int job_run_Id { get; set; }
         public string EffectiveDate { get; set; }
